requirements.txt
Pillow==9.1.0
pyserial==3.5
